+++ neverwinter/nwscript/compilerapi.cpp
@@ -1,0 +1,39 @@
+#include "native/exobase.h"
+#include "native/scriptcomp.h"
+
+extern "C" CScriptCompiler* scriptCompApiNewCompiler(
+    char* lang, int src, int bin, int dbg,
+    int32_t (*ResManWriteToFile)(const char* sFileName, RESTYPE nResType, const uint8_t* pData, size_t nSize, bool bBinary),
+    const char* (*ResManLoadScriptSourceFile)(const char* fn, RESTYPE rt),
+    const char* (*TlkResolve)(STRREF strRef),
+    bool writeDebug
+)
+{
+    CScriptCompilerAPI api;
+    api.ResManUpdateResourceDirectory = +[](const char* sAlias) -> BOOL { return FALSE; };
+    api.ResManWriteToFile = ResManWriteToFile;
+    api.ResManLoadScriptSourceFile = ResManLoadScriptSourceFile;
+    api.TlkResolve = TlkResolve;
+    CScriptCompiler* instance = new CScriptCompiler(src, bin, dbg, api);
+    instance->SetCompileSymbolicOutput(0);
+    instance->SetGenerateDebuggerOutput(writeDebug);
+    instance->SetOptimizeBinaryCodeLength(0);
+    instance->SetCompileConditionalOrMain(1);
+    instance->SetIdentifierSpecification(lang);
+    instance->SetOutputAlias("scriptout");
+    return instance;
+}
+
+struct CompileResult
+{
+    int32_t code;
+    char* str; // static buffer
+};
+
+extern "C" CompileResult scriptCompApiCompileFile(CScriptCompiler* instance, char* filename)
+{
+    CompileResult ret;
+    ret.code = instance->CompileFile(filename);
+    ret.str = ret.code ? instance->GetCapturedError()->CStr() : (char*)"";
+    return ret;
+}